requirements.txt
google-cloud-storage==1.20.0
pandas==1.3.5
scikit-learn==1.0.2
numpy==1.22.3
xgboost==1.1.1
pyarrow==8.0.0
Flask==1.1.2
